refactor: reexport config versioned config types
This is part of the migration to Figment in the configuration.

This expose new versioned types (version 1). However, those types still
used the old Config crate. Replacement by Figment has not been done yet.
